Remove unnecessary includes. Remove using directives from .h files. Use cstdint instead of boost's.
git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@3406 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534

+++ cpp/src/libxtreemfs/vivaldi.cpp
@@ -24,6 +24,7 @@
 #include "libxtreemfs/xtreemfs_exception.h"
 #include "util/logging.h"
 #include "util/zipf_generator.h"
+#include "xtreemfs/DIRServiceClient.h"
 #include "xtreemfs/GlobalTypes.pb.h"
 #include "xtreemfs/OSDServiceClient.h"
 
@@ -345,11 +346,11 @@ void Vivaldi::Run() {
       vivaldi_iterations = (vivaldi_iterations + 1) % LONG_MAX;
 
       // Sleep until the next iteration
-      uint64_t sleep_in_s = static_cast<uint64_t>(
+      uint32_t sleep_in_s = static_cast<uint32_t>(
           vivaldi_options_.vivaldi_recalculation_interval_s -
           vivaldi_options_.vivaldi_recalculation_epsilon_s +
           (static_cast<double>(rand()) / (RAND_MAX - 1)) *
-          2.0 * vivaldi_options_.vivaldi_recalculation_epsilon_s);
+           2.0 * vivaldi_options_.vivaldi_recalculation_epsilon_s);
 
       if (Logging::log->loggingActive(LEVEL_DEBUG)) {
         Logging::log->getLog(LEVEL_DEBUG)
